// app/src/generated/models/Al_sectionsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Al_sectionsal_ownerrole = {
  120910100: 'Tax team',
  120910101: 'AQS checker',
  120910102: 'Adviser',
  120910103: 'T&C Manager',
  120910104: 'Manager / Admin'
} as const;
export type Al_sectionsal_ownerrole = keyof typeof Al_sectionsal_ownerrole;
export const Al_sectionsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Al_sectionsstatecode = keyof typeof Al_sectionsstatecode;
export const Al_sectionsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Al_sectionsstatuscode = keyof typeof Al_sectionsstatuscode;

export interface Al_sectionsBase {
  "al_checklistversionid@odata.bind": string;
  al_displayorder?: number;
  al_helptext?: string;
  al_isconditional: boolean;
  al_name: string;
  al_ownerrole: Al_sectionsal_ownerrole;
  al_sectioncode: string;
  al_sectionid: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  statecode: Al_sectionsstatecode;
  statuscode?: Al_sectionsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Al_sections extends Al_sectionsBase {
  al_checklistversionidname?: string;
  al_isconditionalname?: string;
  al_ownerrolename?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  organizationidname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  al_checklistversionid?: object;
  _al_checklistversionid_value?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  organizationid?: object;
  _organizationid_value?: string;
}