// app/src/generated/models/Al_questionsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Al_questionsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Al_questionsstatecode = keyof typeof Al_questionsstatecode;
export const Al_questionsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Al_questionsstatuscode = keyof typeof Al_questionsstatuscode;

export interface Al_questionsBase {
  al_displayorder?: number;
  al_name: string;
  al_questioncode: string;
  al_questionid: string;
  "al_sectionid@odata.bind": string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  statecode: Al_questionsstatecode;
  statuscode?: Al_questionsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Al_questions extends Al_questionsBase {
  al_sectionidname?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  organizationidname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  al_sectionid?: object;
  _al_sectionid_value?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  organizationid?: object;
  _organizationid_value?: string;
}